[PR 13374] Controller dialog: Keyboard mapping support
This adds keyboard keys as valid controller features in the controller
dialog. Symbols are used to identify keys in the buttonmap.

A "select key" button is placed above the keys so that the user can
select which key to map by typing the key on their keyboard. This beats
scrolling through a long list.